+++ cartoframes/assets/src/legends.js
