+++ cartoframes/assets/src/widgets.js
