+++ cartoframes/assets/style/themes/light.html.j2
