+++ cartoframes/assets/style/widgets.html.j2
