+++ cartoframes/assets/templates/viz/basic.html.j2
@@ -1,4 +1,4 @@
-{% extends "iframe.html.j2" %}
+{% extends "templates/iframe.html.j2" %}
 {% block srcDoc %}
-  {% include 'viz/main.html.j2' %}
+  {% include 'templates/viz/main.html.j2' %}
 {% endblock %}
