--- cartoframes/assets/templates/viz/init.js.j2
@@ -1,27 +1,0 @@
-const responsive = document.querySelector('as-responsive-content');
-
-function setReady () {
-  const show_info = '{{show_info}}' === 'True';
-  const layers = {{ layers|tojson }};
-  const bounds = {{ bounds }};
-  const camera = null;
-  const mapboxtoken = '{{mapboxtoken}}';
-  const basemap = '{{basemap}}';
-  const basecolor = '{{basecolor}}';
-  const has_legends = '{{has_legends}}' === 'True';
-  const default_legend = '{{default_legend}}' === 'True';
-
-  init({
-    show_info,
-    layers,
-    bounds,
-    camera,
-    mapboxtoken,
-    basemap,
-    basecolor,
-    has_legends,
-    default_legend
-  })
-}
-
-responsive.addEventListener('ready', setReady);
