+++ cartoframes/assets/templates/viz/main.html.j2
@@ -33,18 +33,18 @@
 <body class="as-app-body as-app">
   <as-responsive-content>
     {% if has_widgets %}
-      {% include 'viz/widgets.html.j2' %}
+      {% include 'templates/viz/widgets.html.j2' %}
     {% endif %}
     <main class="as-main">
       <div class="as-map-area">
         <div id="map"></div>
         {% if show_info %}
-          {% include 'viz/info.html.j2' %}
+          {% include 'templates/viz/info.html.j2' %}
         {% endif %}
         {% if has_legends %}
           <div class="as-map-panels" data-name="Legends">
             <div class="as-panel as-panel--left as-panel--top">
-              {% include 'viz/legends.html.j2' %}
+              {% include 'templates/viz/legends.html.j2' %}
             </div> <!-- as-panel -->
           </div> <!-- as-map-panels -->
         {% endif %}
@@ -53,10 +53,10 @@
   </as-responsive-content>
 
   {% if is_embed %}
-    {% include 'viz/footer.html.j2' %}
+    {% include 'templates/viz/footer.html.j2' %}
   {% endif %}
 
-  {% include 'error/basic.html.j2' %}
+  {% include 'templates/error/basic.html.j2' %}
 </body>
 <script>
   {% include 'utils/base64.js.j2' %}
@@ -65,6 +65,6 @@
   {% include 'src/bundle.js' %}
 </script>
 <script>
-  {% include 'viz/init.js.j2' %}
+  {% include 'init.js.j2' %}
 </script>
 </html>
