+++ cartoframes/assets/utils/base64.js.j2
