+++ rollup.config.js
@@ -1,21 +1,21 @@
 import resolve from 'rollup-plugin-node-resolve';
 
 module.exports = {
-  input: 'cartoframes/assets/templates/src/index.js',
+  input: 'cartoframes/assets/src/index.js',
   output: {
-    file: 'cartoframes/assets/templates/src/bundle.js',
+    file: 'cartoframes/assets/src/bundle.js',
     format: 'iife',
     name: 'init'
   },
   plugins: [ resolve() ],
   watch: {
     include: [
-      'cartoframes/assets/templates/src/**/*.js',
-      'cartoframes/assets/templates/**/*.j2'
+      'cartoframes/assets/src/**/*.js',
+      'cartoframes/assets/**/*.j2'
     ],
     exclude: [
       'node_modules/**',
-      'cartoframes/assets/templates/src/bundle.js'
+      'cartoframes/assets/src/bundle.js'
     ]
   }
 };
